Refactor weather data classes and JSON serialization
Removed `using System.Collections.Generic;` from `WeatherData.cs` and moved `CurrentWeather` and `HourlyData` classes to their own files. Added properties for `Windspeed`, `WindSpeed10m`, and `Time` with required lists. Included `JsonPropertyName` attributes for proper JSON serialization.
