+++ ProjectBotenReservering.Core/Models/CurrentWeather.cs
@@ -1,0 +1,10 @@
+﻿using System.Text.Json.Serialization;
+
+namespace ProjectBotenReservering.Core.Models
+{
+    public class CurrentWeather
+    {
+        [JsonPropertyName("windspeed")]
+        public decimal Windspeed { get; set; }
+    }
+}
