+++ ProjectBotenReservering.Core/Models/WeatherData.cs
@@ -1,5 +1,4 @@
 ﻿using System.Text.Json.Serialization;
-using System.Collections.Generic;
 
 namespace ProjectBotenReservering.Core.Models
 {
@@ -10,21 +9,5 @@ namespace ProjectBotenReservering.Core.Models
 
         [JsonPropertyName("hourly")]
         public required HourlyData Hourly { get; set; }
-    }
-
-    public class CurrentWeather
-    {
-        [JsonPropertyName("windspeed")]
-        public decimal Windspeed { get; set; }
-    }
-
-    public class HourlyData
-    {
-        [JsonPropertyName("wind_speed_10m")]
-        public required List<decimal> WindSpeed10m { get; set; }
-
-        // Als je ook tijden wilt opslaan:
-        [JsonPropertyName("time")]
-        public required List<string> Time { get; set; }
     }
 }
